Bugfix: f(x) => 0 * x is not invertible.
The original implemention of ConversionFactorFunction returned f(x) when asked for the inverse of f(x) where f(x) => 0 * x. However, this is not an invertible function - there is no function such that f(0*x) = x. Fixed by adding a NotInvertibleException and throwing this exception if this corner case is encountered.